requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.0
blinker==1.9.0
certifi==2020.6.20
cffi==1.17.1
click==8.1.8
cryptography==3.4.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.0
greenlet==3.1.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==9.2.0
pillow==11.1.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
qrcode==8.0
six==1.17.0
SQLAlchemy==2.0.27
typing_extensions==4.12.2
Werkzeug==3.1.0
WTForms==3.1.2
